+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 2.3.0 - Unreleased
 
  * Add support to trigger on specific ip address
- * Add variable feature which allow to use request properties in action values
+ * Add variable feature which allow using request properties in action values
+ * Add append/prepend/replace text content
  * Fix cases when an url could not be parsed
  * Fix wrong rule ids reported in log or header